+++ Sources/UICollectionViewKit/Protocols/ItemDisplayable.swift
@@ -4,4 +4,9 @@ public protocol ItemDisplayable: Hashable {
     var itemID: String { get }
     var categoryID: String { get }
     var imageURL: URL { get }
+    var isAnimatedWebP: Bool { get }
+}
+
+public extension ItemDisplayable {
+    var isAnimatedWebP: Bool { false }
 }